eorlive_v2/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Script
Flask-Login
Flask-Cache
